# 8267687: ModXNode::Ideal optimization is better than Parse::do_irem