--- a/test/micro/org/openjdk/bench/vm/compiler/ModPowerOf2.java
+++ b/test/micro/org/openjdk/bench/vm/compiler/ModPowerOf2.java
@@ -19,7 +19,7 @@
  * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
  * or visit www.oracle.com if you need additional information or have any
  * questions.
- * 
+ *
  */
 package org.openjdk.bench.vm.compiler;