@kelthuzadx 2021-06-01
Hi Nils, your concerns are very reasonable! I will add more benchmark results on different archs later(This may take a while).