@vnkozlov 2021-06-04
I have question about benchmarks results. The unit is `ns/op` - the less number is better. And `Parse::do_irem` gives smaller numbers. Why then `Ideal` is better?
I understand that logically code without branch should be faster. But if the branch is predicted by CPU it may be not the case.
I am for to have only one place to do the optimization. But I want to see correct numbers which we can based on.
Also would be nice to see numbers for mixed (negative and positive) loop.