@kelthuzadx 2021-06-04
Hi Vladimir,

> I have question about benchmarks results. The unit is `ns/op` - the less number is better. And `Parse::do_irem` gives smaller numbers. Why then `Ideal` is better?

Thanks for your reminder, I overlook the unit `ns/op`.

> I understand that logically code without branch should be faster. But if the branch is predicted by CPU it may be not the case.
> I am for to have only one place to do the optimization. But I want to see correct numbers which we can based on.
> Also would be nice to see numbers for mixed (negative and positive) loop.

It seems test is indeed affected by the branch prediction. When I add mixed computations loop, ModINode::Ideal is better than Parse::do_irem

```
x64
-----
Parse::do_irem
Benchmark                         Mode  Cnt      Score    Error  Units
ModPowerOf2.testMixedPowerOf2     avgt   25  13258.945 ± 12.420  ns/op
ModPowerOf2.testNegativePowerOf2  avgt   25   8685.782 ± 16.010  ns/op
ModPowerOf2.testPositivePowerOf2  avgt   25   6604.848 ±  1.402  ns/op

ModINode::Ideal
Benchmark                         Mode  Cnt      Score    Error  Units
ModPowerOf2.testMixedPowerOf2     avgt   25  12918.206 ± 11.730  ns/op
ModPowerOf2.testNegativePowerOf2  avgt   25   8675.248 ±  3.279  ns/op
ModPowerOf2.testPositivePowerOf2  avgt   25   8678.828 ±  6.336  ns/op

AArch64:
------
Parse::do_irem
Benchmark                         Mode  Cnt      Score   Error  Units
ModPowerOf2.testMixedPowerOf2     avgt   25  12340.268 ± 1.015  ns/op
ModPowerOf2.testNegativePowerOf2  avgt   25   6752.467 ± 2.671  ns/op
ModPowerOf2.testPositivePowerOf2  avgt   25   6545.322 ± 0.666  ns/op

ModINode::Ideal
Benchmark                         Mode  Cnt      Score   Error  Units
ModPowerOf2.testMixedPowerOf2     avgt   25  12339.301 ± 1.406  ns/op
ModPowerOf2.testNegativePowerOf2  avgt   25   6753.852 ± 3.460  ns/op
ModPowerOf2.testPositivePowerOf2  avgt   25   6752.895 ± 4.207  ns/op
```

So I must correct my previous conclusion. On aarch64, there is no obvious difference between these two optimizations. On x64, Parse::do_irem is better than ModINode::Ideal w/ branch prediction, otherwise ModINode::Ideal is a better choice.
[new_bench_aarch64.log](https://github.com/openjdk/jdk/files/6595966/new_bench_aarch64.log)
[new_bench_x64.log](https://github.com/openjdk/jdk/files/6595967/new_bench_x64.log)